[HW-06] MainActivity ViewModel 수정
